+++ CHANGELOG.md
@@ -8,6 +8,14 @@
 <b>Version 1.1.90</b> - October 2020<br/>
 - Connection and queue handling optimization for big installations where knxd is used and where there is a near maximum (allowed by knx standards) of datagram per seconds traffic on the BUS. Thanks @Songzh<br/>
 - FIX sample message not showing in the config window of knx-ultimate device, if the device is already presento on the flow. <br/>
+- Added sample for Datapoint 2.*, 1 bit with priority. <br/>
+- Added sample for Datapoint 6.x, value -128 to 127%. <br/>
+- Added sample for Datapoint 7.x. <br/>
+- Added sample for Datapoint 8.x. <br/>
+- Added sample for Datapoint 9.x. <br/>
+- Added sample for Datapoint 12.x. <br/>
+- Added sample for Datapoint 13.x. <br/>
+- Added sample for Datapoint 14.x. <br/>
 </p>
 <p>
 <b>Version 1.1.89</b> - September 2020<br/>